+++ src/test/java/info/jab/jbang/InitCommandTest.java
@@ -6,6 +6,7 @@ import info.jab.jbang.behaviours.EditorConfig;
 import info.jab.jbang.behaviours.GithubAction;
 import info.jab.jbang.behaviours.Maven;
 import info.jab.jbang.behaviours.QuarkusCli;
+import info.jab.jbang.behaviours.Sdkman;
 import info.jab.jbang.behaviours.SpringCli;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
@@ -48,6 +49,9 @@ class InitCommandTest {
     @Mock
     private EditorConfig mockEditorConfig;
 
+    @Mock
+    private Sdkman mockSdkman;
+
     private InitCommand initCommand;
     private final ByteArrayOutputStream outputStreamCaptor = new ByteArrayOutputStream();
     private final PrintStream originalOut = System.out;
@@ -62,7 +66,8 @@ class InitCommandTest {
                 mockQuarkusCli,
                 mockCursor,
                 mockGithubAction,
-                mockEditorConfig
+                mockEditorConfig,
+                mockSdkman
         );
 
         // Capture console output for assertions